修复权限问题
