+++ src/common/paraparse/host.go
@@ -1,15 +1,15 @@
 /*
  * Tencent is pleased to support the open source community by making 蓝鲸 available.
  * Copyright (C) 2017-2018 THL A29 Limited, a Tencent company. All rights reserved.
- * Licensed under the MIT License (the "License"); you may not use this file except 
+ * Licensed under the MIT License (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
  * http://opensource.org/licenses/MIT
  * Unless required by applicable law or agreed to in writing, software distributed under
  * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language governing permissions and 
+ * either express or implied. See the License for the specific language governing permissions and
  * limitations under the License.
  */
- 
+
 package params
 
 import (
@@ -80,6 +80,16 @@ func ParseHostParams(input []interface{}, output map[string]interface{}) error {
 		case common.BKDBIN:
 			d := make(map[string]interface{})
 			d[operator] = value
+			output[field] = d
+		case common.BKDBLIKE:
+			d := make(map[string]interface{})
+			valStr, ok := value.(string)
+			if ok {
+				d[operator] = SpeceialCharChange(valStr)
+			} else {
+				d[operator] = value
+			}
+
 			output[field] = d
 		default:
 			d := make(map[string]interface{})
