+++ src/web_server/application/middleware/validaccess.go
@@ -17,102 +17,11 @@ import (
 	"configcenter/src/common/blog"
 	"configcenter/src/common/util"
 	"encoding/json"
-	"regexp"
 	"strings"
 
 	"github.com/gin-gonic/contrib/sessions"
 	"github.com/gin-gonic/gin"
 )
-
-const BK_CC_MAINTAINERS string = "bk_biz_maintainer"
-
-const (
-	BK_CC_EVENT    string = "event"
-	BK_CC_MODEL    string = "model"
-	BK_CC_AUDIT    string = "audit"
-	BK_CC_RESOURCE string = "resource"
-)
-
-const (
-	BK_CC_UPDATE     string = "update"
-	BK_CC_DELETE     string = "delete"
-	BK_CC_CREATE     string = "create"
-	BK_CC_SEARCH     string = "search"
-	BK_CC_HOSTUPDATE string = "hostupdate"
-	BK_CC_HOSTTRANS  string = "hosttrans"
-	BK_CC_TOPOUPDATE string = "topoupdate"
-	BK_CC_CUSTOMAPI  string = "customapi"
-	BK_CC_PROCCONFIG string = "proconfig"
-)
-
-//bk search
-const (
-	BK_APP_SEARCH    = "biz/search"
-	BK_SET_SEARCH    = "set/search"
-	BK_MODULE_SEARCH = "module/search"
-	BK_INST_SEARCH   = "inst/search"
-	BK_HOSTS_SEARCH  = "hosts/search"
-	BK_HOSTS_SNAP    = "hosts/snapshot"
-	BK_HOSTS_HIS     = "hosts/history"
-	BK_TOPO_MODEL    = "topo/model"
-)
-
-//bk topo
-const (
-	BK_INSTS  string = "inst"
-	BK_TOPO   string = "topo"
-	BK_SET    string = "set"
-	BK_MODULE string = "module"
-)
-
-const BK_CC_CLASSIFIC string = "object/classifications"
-
-const BK_CC_OBJECT_ATTR string = "object/attr/search"
-
-const BK_CC_PRIVI_PATTERN string = "topo/privilege/user/detail"
-
-const BK_OBJECT_ATT_GROUP = "objectatt/group/property/owner"
-
-const BK_CC_USER_CUSTOM string = "usercustom"
-
-const BK_CC_HOST_FAVORITES string = "favorites"
-
-//proc manage pattern
-const BK_PROC_S string = "proc"
-
-//user api pattern
-const BK_USER_API_S string = "userapi"
-
-//host trans pattern
-const BK_HOST_TRANS = "hosts/modules"
-
-//system config privilege pattern
-const (
-	resPattern    = `(hosts/import|export)|(hosts/modules/resource/idle)`
-	objectPattern = `object/classification/[a-z0-9A-Z]+/objects$`
-)
-
-//system config privilege regexp
-var (
-	resPatternRegexp    = regexp.MustCompile(resPattern)
-	objectPatternRegexp = regexp.MustCompile(objectPattern)
-)
-
-//host update string
-const BK_HOST_UPDATE string = "hosts/batch"
-
-//host manage search pattern
-const searchPattern = `(hosts/[\w]+/[\d]+)|(topo/internal/[\w]+/[\w]+)|(topo/inst/[\w]+/[\d]+)|(object/host/inst/[\d]+)`
-
-var searchPatternRegexp = regexp.MustCompile(searchPattern)
-
-var BK_CC_SYSCONFIG = []string{"event", "model", "audit"}
-
-var BK_CC_MODEL_PRE = []string{"object", "objects"}
-
-var BK_CC_AUDIT_PRE = []string{"audit"}
-
-var BK_CC_EVENT_PRE = []string{"event"}
 
 //ValidResAccess valid resource access privilege
 func ValidResAccess(pathArr []string, c *gin.Context) bool {
@@ -199,6 +108,22 @@ func ValidResAccess(pathArr []string, c *gin.Context) bool {
 			}
 			return validModelConfigPrivi(modelPrivi, method, pathArr)
 		}
+		return false
+
+	}
+
+	//valid inst import privilege
+	if strings.Contains(pathStr, BK_INSTSI) && !strings.Contains(pathStr, BK_IMPORT) {
+		est := c.GetHeader(common.BKAppIDField)
+		if "" == est {
+			modelPrivi := session.Get("modelPrivi").(string)
+			if 0 == len(modelPrivi) {
+				blog.Error("get model privilege json error")
+				return false
+			}
+			return validInstsOpPrivi(modelPrivi, method, pathArr)
+		}
+		return false
 
 	}
 
@@ -265,7 +190,33 @@ func validSysConfigPrivi(sysPrivi interface{}, config string) bool {
 	return false
 }
 
-//validModelConfigPrivi valid model inst privilege
+//validInstsOpPrivi  valid inst operation privilege
+func validInstsOpPrivi(modelPrivi, method string, pathArr []string) bool {
+	var mPrivi map[string][]string
+	var objName string
+	err := json.Unmarshal([]byte(modelPrivi), &mPrivi)
+	if nil != err {
+		blog.Error("get model privilege json error")
+		return false
+	}
+	objName = pathArr[len(pathArr)-2]
+	priviArr, ok := mPrivi[objName]
+	if false == ok {
+		blog.Error("get object privilege  error")
+		return false
+	}
+	if util.InArray(BK_CC_UPDATE, priviArr) {
+		return true
+	}
+	if util.InArray(BK_CC_CREATE, priviArr) && !util.InArray(BK_CC_SEARCH, pathArr) {
+		return true
+	}
+
+	blog.Error("inst op privilege valid not pass")
+	return false
+}
+
+// validModelConfigPrivi valid model inst privilege
 func validModelConfigPrivi(modelPrivi string, method string, pathArr []string) bool {
 
 	var mPrivi map[string][]string
@@ -286,9 +237,16 @@ func validModelConfigPrivi(modelPrivi string, method string, pathArr []string) b
 		blog.Error("get object privilege  error")
 		return false
 	}
-	//valid update privilege
-	if method == common.HTTPUpdate && util.InArray(BK_CC_UPDATE, priviArr) {
-		return true
+
+	//merge update&&create privilege
+	if method == common.HTTPUpdate || method == common.HTTPCreate {
+		if util.InArray(BK_CC_UPDATE, priviArr) {
+			return true
+		}
+		if util.InArray(BK_CC_CREATE, priviArr) && !util.InArray(BK_CC_SEARCH, pathArr) {
+			return true
+		}
+
 	}
 
 	//valid delete privilege
@@ -296,16 +254,11 @@ func validModelConfigPrivi(modelPrivi string, method string, pathArr []string) b
 		return true
 	}
 
-	//valid create privilege
-	if method == common.HTTPCreate && util.InArray(BK_CC_CREATE, priviArr) && !util.InArray(BK_CC_SEARCH, pathArr) {
-		return true
-	}
-
 	//valid search privilege
 	if method == common.HTTPSelectPost && util.InArray(BK_CC_SEARCH, priviArr) && util.InArray(BK_CC_SEARCH, pathArr) {
 		return true
 	}
-	blog.Error("modle privilege valid not pass")
+	blog.Error("model privilege valid not pass")
 	return false
 }
